requirements.txt
azure-cosmos
azure-functions
numpy
pandas
python-dotenv